update testsuites, support gettimeofday and pipe, correct errors of wait

+++ codes/os/src/syscall/fs.rs
@@ -162,7 +162,7 @@ pub fn sys_close(fd: usize) -> isize {
     0
 }
 
-pub fn sys_pipe(pipe: *mut usize) -> isize {
+pub fn sys_pipe(pipe: *mut u32) -> isize {
     let task = current_task().unwrap();
     let token = current_user_token();
     let mut inner = task.acquire_inner_lock();
@@ -171,8 +171,8 @@ pub fn sys_pipe(pipe: *mut usize) -> isize {
     inner.fd_table[read_fd] = Some(FileClass::Abstr(pipe_read));
     let write_fd = inner.alloc_fd();
     inner.fd_table[write_fd] = Some(FileClass::Abstr(pipe_write));
-    *translated_refmut(token, pipe) = read_fd;
-    *translated_refmut(token, unsafe { pipe.add(1) }) = write_fd;
+    *translated_refmut(token, pipe) = read_fd as u32;
+    *translated_refmut(token, unsafe { pipe.add(1) }) = write_fd as u32;
     0
 }
